picture overview and renaming
