+++ src/zerodose/workflows/__init__.py
@@ -1,17 +1,17 @@
 """Pipeline for ZeroDose project."""
 from .abnormality import create_abnormality_maps
-from .full import run_with_registration
-from .main import run
-from .niftyreg import from_mni
-from .niftyreg import to_mni
 from .normalization import normalize_to_pet
+from .pipeline import pipeline
+from .registration import from_mni
+from .registration import to_mni
+from .run import run
 from .synthetization import synthesize_baselines
 
 
 __all__ = [
     "create_abnormality_maps",
     "run",
-    "run_with_registration",
+    "pipeline",
     "from_mni",
     "to_mni",
     "normalize_to_pet",
